Drop CMAKE_SYSTEM_VERSION from modules json if Linux
In order to make the project build reproducible it should not record
transient information of the build host.

The linux kernel is not encoded in the source, and the value from the
modules.json does not seem to be used anywhere, so we can drop it for
now.

Fixes: QTBUG-136060
Change-Id: Ic157d66d47040b080e9121fd065a8d60d489f586
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>
(cherry picked from commit dc52aa68ea305d8ef40f8b514bfeba80cd6c2c1e)